+++ packages/podlite-schema/src/plugin-formatting-codes.ts
@@ -34,9 +34,10 @@ const middle: ParserPlugin = () => tree => {
       const conf = makeAttrs(n, ctx)
       const isCodeBlock = 'name' in n && n.name === 'code'
       const isDataBlock = 'name' in n && n.name === 'data'
+      const isMarkdownBlock = 'name' in n && n.name === 'markdown'
       const allowValues = conf.getAllValues('allow')
       // for code block not parse content by default
-      if ((isCodeBlock || isDataBlock) && allowValues.length == 0) return n
+      if ((isCodeBlock || isDataBlock || isMarkdownBlock) && allowValues.length == 0) return n
       const allowed = allowValues.sort()
 
       const transformer = makeTransformer({